+++ Changelog.md
@@ -8,6 +8,7 @@
 * `DataType::display_size` now returns `Option<NonZeroUSize>` instead of `Option<usize>`.
 * `DataType::utf8_len` now returns `Option<NonZeroUSize>` instead of `Option<usize>`.
 * `DataType::utf16_len` now returns `Option<NonZeroUSize>` instead of `Option<usize>`
+* `DataType::column_size` now returns `Option<NonZeroUSize>` instead of `usize`
 * `BufferDesc::from_data_type` now returns `None` for variadic types without upper bound instead of a zero sized buffer.
 
 ## 3.0.1